requirements.txt
python-dotenv == 0.21.0
pandas == 1.5.1
robin-stocks == 2.1.0
boto3 == 1.25.2
polygon-api-client == 1.3.1
pytz == 2022.5
vectorbt == 0.24.2
scipy == 1.9.3
scikit-learn == 0.24.2
auto-sklearn == 0.14.6
cryptography == 38.0.1
ta == 0.10.2
python-binance == 1.0.16
imbalanced-learn == 0.8.1
icosphere == 0.1.3
pynisher == 0.6.4
